backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
pydantic
requests
twilio
aiofiles